moves linkstatistic to researches
