+++ js/analyzer.js
@@ -10,7 +10,7 @@ var findKeywordInFirstParagraph = require( "./analyses/findKeywordInFirstParagra
 var findKeywordInPageTitle = require( "./analyses/findKeywordInPageTitle.js" );
 var getKeywordDensity = require( "./analyses/getKeywordDensity.js" );
 var countImages = require( "./analyses/getImageStatistics.js" );
-var countLinks = require( "./analyses/getLinkStatistics.js" );
+var countLinks = require( "./researches/getLinkStatistics.js" );
 var getKeyphraseLength = require( "./analyses/getWordCount.js" );
 var isUrlTooLong = require( "./analyses/isUrlTooLong.js" );
 var getSubheadings = require( "./analyses/matchKeywordInSubheadings.js" );
