+++ js/researcher.js
@@ -8,7 +8,7 @@ var isEmpty = require( "lodash/lang/isEmpty" );
 // assessments
 var wordCountInText = require( "./researches/wordCountInText.js" );
 var findKeywordInPageTitle = require( "./analyses/findKeywordInPageTitle.js" );
-var getLinkStatistics = require( "./analyses/getLinkStatistics.js" );
+var getLinkStatistics = require( "./researches/getLinkStatistics.js" );
 var calculateFleschReading = require( "./researches/calculateFleschReading" );
 
 /**
