+++ js/researches/getLinkStatistics.js
